pyproject.toml
[tool.poetry]
name = "lms-scypro"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "lms_scypro"}]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.5"
pillow = "^10.0.1"
psycopg2 = "^2.9.7"
djangorestframework = "^3.14.0"
django-filter = "^23.3"
djangorestframework-simplejwt = "^5.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
